Separate core types into stable crate
This change moves the fundamental message-related traits and types into
their own hedwig_core crate. This crate will have more version stability
than the primary crate, which should allow less dependency churn when a
set of message types are used across many crates.

Further divisions of the library were considered. For example, it could
be sensible to separate out hedwig-gcp, hedwig-prost, hedwig-json, etc.
This would enable even better upgrade granularity, but at a maintenance
cost. Both this repo and its users would then have to
specify/coordinate/upgrade a plethora of crates to get a working system.
Ultimately a smaller split of core+rest seems like a better compromise.
