+++ hedwig_core/src/topic.rs
